Draft: move svg subfunctions to svgtext.py module
